Isolated frame creation
